major code refactor
