+++ riemann/__init__.py
@@ -1,6 +1,13 @@
-from .riemann import Sampler, Model, Proposal, ParameterError
-from . import proposals
-from . import models
-from . import samplers
-
-__all__ = ['Sampler', 'Model', 'Proposal', 'proposals', 'models']
+# from .riemann import ParameterError
+# from riemann.samplers.sampler import Sampler
+# from riemann.proposals.proposal import Proposal
+# from riemann.models.model import Model
+# from . import proposals
+# from . import models
+# from . import samplers
+#
+# __all__ = ['proposals', 'models']
+from riemann.models.model import Model
+from riemann.samplers.sampler import Sampler
+from riemann.proposals.proposal import Proposal
+from riemann.sampling_errors import *
