+++ riemann/models/__init__.py
@@ -1,1 +1,2 @@
-from .gaussian import MultiGaussianDist, MultiGaussianModel, MixtureDist
+# from .gaussian import MultiGaussianDist, MultiGaussianModel, MixtureDist
+# from riemann.models.model import Model
