+++ riemann/proposals/__init__.py
@@ -1,1 +1,1 @@
-from .randomwalk import MetropolisRandomWalk, AdaptiveMetropolisRandomWalk, pCN
+# from .randomwalk import MetropolisRandomWalk, AdaptiveMetropolisRandomWalk, pCN
