+++ riemann/samplers/__init__.py
@@ -1,1 +1,1 @@
-from .ptsampler import PTSampler
+# from .ptsampler import PTSampler
